Create Aujasvit and the Circle Game.cpp
Aujasvit just came up with a new game to play with his friends. N people stand in a circle, each assigned an index from 1 to N in clockwise order. Then the following operation is done N−1

times.

    The person with the lowest index holds a coin.
    Then, the coin then moves to the next person clockwise M−1

    times.
    Then, the person who is holding the coin is removed from the circle.

The last person remaining after N−1

operations is the winner.

Aujasvit has already decided on the integer M
to be used in the game. Before inviting people to play his game, Aujasvit wants to know the winner if the game has 1 player, 2 players, 3 players, …, X players. That is, you need to output the winner when N=1,N=2,...,N=X.