Create Chef and Fixed Deposits.cpp
Chef wants to make a purchase. For this, he needs X

gold coins, but he has none at the moment.

Chef has N
fixed deposits, the ith of which is worth Ai coins. He wants to open the minimum number of these deposits so that he has at least X

coins.

You have to tell Chef the minimum number of fixed deposits he must open in order to have X
coins, or say that this is impossible.