Create Crying Colours.cpp
You have a total of 3N balls of colours Red, Blue and Green. There are exactly N

balls of each colour.

These 3N
balls are now distributed to 3 boxes such that each box contains exactly N

balls. You are given the contents of each box.

You would like the 1st
box to contain all the red balls, the 2nd box to contain all the green balls, and the 3rd

box to contain all the blue balls. Note that the given order of boxes is important here — it is not enough for each box to contain only balls of a single colour.

To achieve this, you can perform the following operation several (possibly, zero) times:

    Pick any two distinct boxes, pick any one ball from each of these two boxes, and swap them.

Determine the minimum number of operations required to satisfy the given condition.