Create Power Sum.cpp
You are given a sequence A of N integers such that every element is a non-negative power of 2

.

A sequence is called good if its sum is a non-negative power of 2
. You would like to turn A

into a good sequence.

To achieve this, you can perform the following operation on A

:

    Pick a non-empty subsequence of A

, pick a positive integer X such that X≤230, and multiply every element of this subsequence by X

    .

Find the minimum number of operations required to turn A
into a good sequence. Also, find one sequence of operations which does this. If there are multiple possible answers, you may find any of them.