Create C - The Kth Time Query.cpp
We have a sequence of N numbers: A=(a1​,a2​,…,aN​).
Process the

Q queries explained below.

    Query 

i: You are given a pair of integers (xi​,ki​). Let us look at the elements of A one by one from the beginning: a1​,a2​,… Which element will be the ki​-th occurrence of the number xi​?
Print the index of that element, or −1 if there is no such element.