Update and rename 1.cpp to A. Not Shading.cpp
There is a grid with n rows and m

columns. Some cells are colored black, and the rest of the cells are colored white.

In one operation, you can select some black cell and do exactly one of the following:

    color all cells in its row black, or
    color all cells in its column black. 

You are given two integers r
and c. Find the minimum number of operations required to make the cell in row r and column c black, or determine that it is impossible.