Create Divisors and Reciprocals.cpp
Alice is teaching Bob maths via a game called N

-guesser.

Alice has a positive integer N

which Bob needs to guess. She gives him two pieces of information with which to do this:

    A positive integer X

, which denotes the sum of divisors of N
.
Two positive integers A
and B, which denote that the sum of reciprocals of divisors of N is A/B

    .

Bob either needs to guess N

or tell that no such number exists.

It can be shown that if a valid N
exists, it is unique.