Create Binary Base Basics.cpp
You are given a binary string S and an integer K. In one operation, you can pick any bit and flip it, i.e turn 0 to 1 or 1 to 0. Can you make the string S a palindrome using exactly K

operations?

Print YES if this is possible, and NO if it is not.